refactor: create a separate hook for file upload with the new request
